replaced linking against modified libdbus with a wrapper for the relevant dbus function
The wrapper for dbus_connection_send_with_reply() sets the increased timeout
if necessary, then calls the original implementation via dlsym(RTLD_NEXT). This
ensures that always the right libdbus of the host system is used, which is
necessary for ITOS2008 (which has a version which is incompatible with ITOS2007/06).


git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@488 15ad00c4-1369-45f4-8270-35d70d36bdcd