made detached recurrence handling more resilient
Child items can now be inserted without the
corresponding parent or before the parent.
Semantically this makes no sense, but servers
which do not know about the semantic might do
it. Also removing the parent without the children
is now supported.

Technically this has to workaround limitations of
the EDS API and is done as:
* child before parent: temporarily save and remove children, insert parent, update parent with children
* remove parent, but not child: temporarily save and remove children, reinsert after removal of parent



git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@627 15ad00c4-1369-45f4-8270-35d70d36bdcd