added synchronization of Evolution memo summary
Most devices only synchronize plain text and do
not have a separate summary field. Such an extra
summary field was added to Evolution after memo
support was initially implemented in SyncEvolution,
therefore SyncEvolution did not transmit that field.

Added transmitting the summary by inserting it
as first line of the plain text blob *if* it is
not already identical with the first line. When
receiving a memo, the summary is set from the first
line *without* removing the first line because
the first line might have been used as a normal
part of the memo.


git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@631 15ad00c4-1369-45f4-8270-35d70d36bdcd