SF #1991286: when creating config for the first time, only enable sync sources which can be synchronized
The check for that was completely missing. Now SyncEvolution checks that the backend is available and
there is at least one database (the first one will be synchronized unless explicitly changed). If these
checks fail and the sync source was explicitly requested by the user by listing it after the server
name, then an error is printed and no configuration is written. If the user wants the default setup,
then the source is silently disabled.


git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@650 15ad00c4-1369-45f4-8270-35d70d36bdcd