improved command line and its documented
More explicit description of the different operations.
Added --run|-r switch: if properties are on the command
line, then an operation must be specified explicitly,
with --run standing for a sync run. Otherwise a
sync might accidentally be done when a configuration
change was intented, but the --configure switch was
forgotten.


git-svn-id: https://zeitsenke.de/svn/SyncEvolution/trunk@696 15ad00c4-1369-45f4-8270-35d70d36bdcd