Synthesis: connect engine with EvolutionSyncSource
SynthesisDBPlugin provides all C calls potentially used by
the Synthesis engine. It needs to be linked into the executable
or libsyncevolution.so so that the SyncEvolution_* calls
are visible to libsynthesis. They will get called when "SyncEvolution"
is specified as plugin name, using dlsym(RTLD_DEFAULT).

The "-uSyncEvolution_Module_Version -Wl,--export-dynamic" flags
are a bit platform specific; a configure check would be useful,
with compilation as a normal plugin as default.

syncevolution.xml is compiled into libsyncevolution. The <datastore>
entries are created for all active sync sources and replace
the <datastore/> place holder in the default configuration.

The configuration can be overridden by placing a syncevolution.xml
file in the current directory, with or without <datastore>
specifications. <datastore/> is replaced as usual.

<dbtypeid> values are generated as hash from the sync source name.
Collisions are not handled yet.