compilation: fixes + installation of client-test + version bump
Removed a spurious package check for synthesis which was reintroduced
during the recent code cleanup and merging.

Fixed compilation with --enable-unit-tests and/or --enable-integration-tests.
CPPUnit header files were not included properly.

When at least one of these enable options is used, client-test is built
as part of the normal "make all" and installed in bindir as part of
"make install". The test files go into the docdir.