HTTP transport: check reply content type
Instead of passing an invalid reply into the Synthesis engine where
it triggers an obscure (WB)XML parser error in the SyncML Toolkit,
check the content type and dump a non-(WB)XML reply into our own log,
then abort with a "transport failed" error.

libsoup already aborted because of the non-200 server status. Now it
also dumps the message in that case.