autotools: first PKG_CHECK_MODULES() must be unconditional
The first PKG_CHECK_MODULES() implicitly checks for the availability of
pkg-config itself. This check is not replicated in later expansions of
this m4 macro. Therefore if the first check is inside an "if ... then ... fi" clause,
then this check is skipped and all following PKG_CHECK_MODULES() calls fail.

autoconf magic - or curse...

As workaround this patch moves the conditional PKG_CHECK_MODULES() further down,
after some unconditional checks. Not sure what the right solution is.