config handling: added removing of a configuration
Only files which are known to belong to SyncEvolution are
deleted. This list is hard-coded in FileConfigTree.cpp and
includes the hidden .synthesis meta files in addition to the
normal config files.

The --remove command line option exposes this new feature.
It removes the configuration *without* asking for confirmation.

Log dirs are not removed yet. This needs to be done via a different
API which will be added as part of access to old session information.