log handling: added access to previous sync sessions
EvolutionSyncClient::getSessions() returns the full pathnames
of the log dirs of the previous sync sessions. --print-sessions
shows those on the command line.

Right now this feature is only useful to remove these directories
or (for the user) to find backup files. Further information has
to written to and extracted from them to make --print-sessions
and its graphical pendant more useful.