log handling: SyncReport is stored in log dir and can be retrieved
Source specific values are stored with a key that is composed of the
prefix "source-" followed by the source name (with - escaped) and
other components representing the various properties. Empty values
are not stored, which is fairly efficient for the statistics array.

When reading back, only sources which have any value are
recreated. Because the sync mode is always set, this ensures
that the set of active sources is restored. Inactive sources
are not part of that set!

--print-sessions now prints the retrieved sync report just like
--sync does. Combined with --quiet only the directory name is printed.