logging: enable log dirs by default
When the "logdir" sync property is unset, then SyncEvolution now
maintains log directories in $XDG_DATA_HOME/applications/syncevolution.
The "maxlogdirs" property defaults to 10 now because of this change.

When the base directory for log dirs ends in "syncevolution", then
the directories created inside it start with the server name, followed
by a date and time (as before). Otherwise "SyncEvolution-" is added.
This is a safety-mechanism that ensures that we have a good chance
of finding the right log directories again with a non-standard
base directory.

The previous default (a single log dir in /tmp with the user name
as part of the log dir name) is no longer supported: when configuring
a non-standard log dir, that directory has to be chosen so that
files from different users aren't mixed up.