added --restore and fixed --status
--status was broken by the last commit: XDG_DATA_DIR was not
expanded. Fixed it as part of the improvements for printing
changes in --restore. Also made some other minor changes
to utility classes as part of these improvements.

The new --restore option can restore data from a data dump,
identified by the directory and a before/after flag. The
corresponding API is EvolutionSyncClient::restore().