ConfigNode + status.ini: added template methods for automatic conversion to/from string
setProperty() and getProperty() now do conversion to/from strings
automatically. There's a specialized implementation for bool which
maps to "true/false" and accepts "true/on/yes" and non-zero values
for true and "false/off/no" for false.

Use this for session status. Fixes a bug where reusing the same
stringstream for formatting accumulated property values: stringstream::clear()
does *not* clear the content, only the error status!

This bug affected -backup-after. It's still not written correctly,
though: the value that gets formatted is always -1. So writing works,
obtaining the value before writing does not.