SyncReport: utility functions for serializing in a ConfigNode
The new >> and << operators convert to resp. from a SyncReport
and a ConfigNode. This is a key/value representation with keys
and value formats which are currently considered SyncEvolution-internal
(and thus not documented). At least they are human-readable, so one
can figure out what they mean.

This is the same format also used for LogDir status.ini files,
which is what the code was written for originally.

The new HashConfigNode can be used in combination with these
utility operators to create an in-memory representation of
a SyncReport with these key/value pairs.