added/improved reporting of local changes
EvolutionSyncClient::getChanges() can be used to query whether
items were changed locally. Added/updated/deleted/total number
of items are reported, plus the time needed to collect that information.
The information is stored in a SyncReport, with ITEM_ANY used
for the total item count.

--status now also does this check and prints the information.

Printing a SyncReport as an ASCII table was considerably
improved for this: columns which are known to be irrelevant
can be suppressed. The layout is a lot more dynamic with
(so I hope) nicer alignment: text related to a source name
is aligned with that name, which itself is flushed to the
right of the name column.