license: changed to LGPL v2.1 + v3
As with the previous change from GPL to LGPL v2.1, this is covered
by copyright ownership and/or the Funambol contributor agreement.

The motivation for adding v3 is greater flexibility regarding
reusing the code in other places and relicensing it.

Also, now the license and disclaimer are mentioned at the start
of the source files. That is not strictly necessary, but more
explicit.