testing: check for correct server behavior during refresh-from-client
During a refresh-from-client sync, the server has to delete all items
that it currently has before adding those sent by the client. This
new test covers that. It complements the testRefreshSemantic test,
which did that for the other direction and thus was renamed to
testRefreshFromClientSemantic.

The server also needs to tell other clients to delete the items
that were deleted this way during a refresh-from-client. This is
not currently tested.