EvolutionMemoSource: notes with only a summary were not sent correctly
Instead of sending the summary, only the empty description was sent.
Now the summary is sent. This has the effect that on a full round-trip,
the summary gets duplicated as part of the description. synccompare
was adapted to not complain about this for the newly added testcase
for this: the script also duplicates the summary.

This cannot be avoided because notes are exchanged as plain text
which cannot distinguish between summary and description.