EvolutionCalendarSource: extra sanity check for empty TZID
The Synthesis engine creates that when receiving a vCalendar 1.0
item; this is considered a bug: http://www.synthesis.ch/indefero/index.php/p/libsynthesis/issues/2/

Without the sanity check, SyncEvolution called e_cal_add_timezone(),
which probably a segfault with some implementations of it (at least
that would explain some of the segfaults). With libecal 2.22 the function
itself didn't crash, but then SyncEvolution itself used the NULL pointer
when trying to report the error.

"skipping VTIMEZONE without TZID" is logged to client.log if this happens now.