build rules: moved D-Bus into sub directories
syncevo-dbus -> src/dbus
interfaces -> src/dbus/interfaces

Files generated from D-Bus interfaces are only built once, in src/dbus/interfaces.
This is changes makes the code layout a bit nicer and, more importantly,
fixes build dependencies when Synthesis is compiled in src.

Now everything that is needed by a certain directory is in that directory
or a sub directory.