configuration handling: only list servers as "configured" which really have a config
Previously, just having a directory was enough to list the server. Now
an additional sanity check ensures that only servers with a configuration
are listed.

This is in response to Bugzilla #1806: removing the config, but leaving
other files led to the situation that --print-servers showed entries
that the other options didn't work with.