logging: logLevel 4 and 5 enable message translation and dumping
At level 4, all messages (regardless whether they were originally
in WBXML or XML format) are translated to XML and logged in that
format (sysynclib_linux_s*_trm*.xml).

At level 5 also the original message is logged (sysync_s*_msg*.[wb]xml).

The level of detail logged to the console is limited to INFO if
a log file is written. Otherwise it would not have been possible to
get a more detailed log file without also increasing the verbosity
of the console output.