config: enabled maxMsgSize/maxObjSize, removed obsolete loSupport (Bugzilla #2784)
The two size options are now copied into the Synthesis XML configuration.
Large object support is always on, so the setting was removed.

Note that the Synthesis engine allocates twice the amount specified in
the configuration and then uses that amount if the server allows it.
That explains why outgoing messages where twice as large as specified.
Incoming messages where inside the size limit.