config: enable clientAuthType (Bugzilla #3242)
This is now mapped to the Synthesis defaultauth config option.
It seems that the option only has an effect during the initial sync
of a client. Later the engine seems to use the setting that was
negotiated with the server during that sync, without the possibility
to override that later on.