config: use ${XDG_CACHE_HOME} for logs and database dumps (Bugzilla #3309)
The reason for this change is that the previous location, XDG_DATA_HOME,
is searched for .desktop files. Creating many files there slows down that
search and thus the Moblin startup.

To fix old installations, the previous directory is moved automatically.
This should work in the majority of the cases. The only known exception
where it fails is when someone explicitly specified the old default
location in the config.ini. In such a setup the directory will always
be moved away even though it is still in use.