libcurl: fixed memory handling issue with libcurl < 7.17.0.
libcurl < 7.17.0 does not copy strings passed into curl_easy_setopt().
As a result, the URL was overwritten with the content type, leading to
a lookup failure. It might also have crashed.

A user on Debian Etch ran into this problem.

To be compatible with a wider range of libcurl versions, the code
now keeps temporary copies of all strings that are passed into
curl_easy_setopt(), regardless whether that is necessary or not.