Debian Etch: fixes some compilation issues
With some tweaks the autotools on Debian Etch are good enough to
compile the code. However, distributions of SyncEvolution should
better be built with recent versions of the tools and then get compiled
on Debian Etch.

We need BOOST >= 1.34 because of BOOST_FOREACH(). The configure script
now checks for this. Debian Etch can get 1.35 from www.backports.org.
That version of BOOST shows that in some files <boost/algorithm/string/join.hpp>
was not explicitly included.