documentation: added vCalendar 1.0 and explain automatic format selection
vCalendar 1.0 is now supported by all backends, but not the default.

In contrast to releases < 0.9, the engine evaluates the DevInf sent
by the server and negotiates the preferred format. The "type" property
can be used to specify the preferred format. Without an explicit
selection, vCard 2.1 and iCalendar 2.0 are the preferred formats.