config + DevInf: always resend DevInf if content of config has changed
Add configdate support to notify the configuration has been changed thus
the Engine will issue a Put command to send the newer version of DevInf.
Changes of configuration are tracked by HashCode hidden configuration.

Without this patch changing the "type" property of a sync source so
that a different version of the datatype was preferred had no effect.
Found when testing with Google, which only works when vCard 2.1 is
marked as preferred.