XML config: generate client config dynamically
Instead of using a static XML file, SyncEvolution uses the XML file
as a template and expands it dynamically in EvolutionSyncClient::getConfigXML().
Information about the device, logging options, datatstores, etc. are all
inserted there.

Empty elements like <debug/> mark the location where certain pieces of
information are supposed to go.

In addition, some of the existing config variables are set in
EvolutionSyncClient::doSync().