XML config: don't encode empty DESCRIPTION in VEVENT/VTODO
The upstream Synthesis config removes a DESCRIPTION if it
is identical with the SUMMARY and then writes an empty DESCRIPTION
property.

SyncEvolution reverts that logic: an empty DESCRIPTION is set
to the SUMMARY. That's more in line with how Evolution works.
Also, libical is unhappy about such empty properties, so don't encode
them if empty (which shouldn't happen, though).