XML config: don't enforce http:// prefix in vCard 2.1 URL
The upstream config declares the WEB field (maps to URL) as "url",
which has the effect that the http:// prefix is added if none
is present.

Evolution does not enforce a particular format nor does it depend on
one. If a user decides to enter a plain string, that's how it should
be synchronized => turned WEB into normal string field.