XML config: be conservative about encoding of EXDATEs
Multiple EXDATEs in the same property are encoded differently
in iCalendar 2.0 and vCalendar 1.0. ScheduleWorld had a problem
with this (bugzilla.moblin.org #2226) for a while. As other
peers might have similar issues, we accept all kinds of
encodings when parsing items and always generate several EXDATE
properties with one value each when encoding.