XML config: use syncclient_sample_config.xml instead of syncevolution.xml
The key difference between the two is that syncclient_sample_config.xml
is kept in sync with the corresponding file on the "synthesis" branch,
which is the unmodified copy of the corresponding file in the libsynthesis
repository.

It is still possible to put an XML file into the current directory
when running SyncEvolution, but now it has to be called
syncclient_sample_config.xml.

Only minor changes are introduced as part of this transition:
- some comment changes
- PERCENT-COMPLETE is now only used in iCalendar 2.0 (as it should be,
  because it is not valid in vCalendar 1.0)
- added vtimezonegenmode and tzidgenmode settings as in upstream
  config (shouldn't change our behavior)