data conversion: different data profiles for SyncML peer + local DB with remote rules
One way to convert data differently on the SyncML and local DB side of the
field list (Bugzilla #2417) is via "remote rules" in the Synthesis engine.
Normally these rules are activated by matching against the peers device information.
This patch adds an rule named "EVOLUTION" which is activated on the Evolution
side by giving its name to the PARSE/MAKETEXTWITHPROFILE() macros.

This depends on an extended Synthesis library.