vCard: support X-MANAGER/ASSISTANT/SPOUSE/ANNIVERSARY (Bugzilla #2418)
These properties are supported in addition to, not instead of, the
traditional X-EVOLUTION- variants. This is done by sending both
property variants to a SyncML server and accepting both when parsing
a vCard. When parsing, X-EVOLUTION- takes precendency over the normal
X- one, which is only relevant in case that they differ.

This is necessary for Funambol, which added support for normal X-
variants in the 7.x server release. ScheduleWorld accepts both
and still works (tested).