compile error: GNOME keyring 2.24 misses extern "C"
Got an undefined reference to C++ GNOME keyring functions
which in reality should have been plain C functions. Reason
was that /usr/include/gnome-keyring-1/gnome-keyring.h (from 2.24.1)
didn't have an 'extern "C"' declarations.

Adding those in our source code fixed that and should be safe,
because double nesting of it is safe and the header is
unlikely to contain real C++ declarations.