WBXML: enabled by default, except for Funambol
Using WBXML reduces message sizes and increases parsing performance.
It was not enabled initially in the 0.9 releases in order to test
this new feature more thoroughly.

Old configs don't have an explicit enableWBXML setting and therefore
will automatically use the new default.

The Funambol server has a known issue with WBXML:
https://forge.ow2.org/tracker/index.php?func=detail&aid=313729&group_id=96&atid=100096
When creating new configs for it, the default is overridden
and WBXML is disabled. Old configs should be updated accordingly,
just to be safe. The bug is unlikely to be triggered, but if it happens,
sync sessions fail completely.