XML config: allow backends to add their own definitions (Bugzilla #2417)
For each of the four entities that a backend might want to add (field
lists, profiles, datatypes, remote rules) a hash is kept which can
be updated by the virtual EvolutionSyncSource methods which are
called while assembling the information for the XML config (getSynthesisInfo(),
getDatastoreXML()).

This information is then subsituted in the right places of the
XML config template, in the right order.