configuration: added "ConsumerReady" option
A property of server template configs. True if the server is
ready for use by "normal" users (everyone can get an account
and some kind of support, we have tested the server well
enough, ...).

Currently the ScheduleWorld and Funambol configs qualify.
Synthesis is not generally available, Memotoo has to be retested
and included in the nightly regression testing.