logging: intercept stdout and stderr (Buzilla #1333)
Added a LogRedirect class which intercepts file descriptors
1 and 2 and channels the data through either a UDP or Unix Domain
datagram socket (UDP being the default, see patch) into the
normal log.

EvolutionSyncClient has to call the parent logger instead of
creating its own LogStdout to get messages to the user.

Currently LogRedirect is not actually used, because it is
a bit too effective (intercepts much of the normal syncevolution
output) and that output is split into different messages (cout
does lots of small writes).