logging: write a single sysync_linux.html log (Bugzilla #3474)
client.log is gone completely. All of the messages written to that
file now also appear in the .html log. The Synthesis engine is told
to write just one log file for global events and the session, using
the new <logsessionstoglobal> option.

The way logging works now is that as soon as possible, SyncEvolution
creates a Synthesis engine without datastores but with correct debug log
settings. Latter that engine is re-initialized with a more complete
XML config which also defines datastores.

The LogDir class implements the SyncEvolution logging base class
and registers itself as the the active logger. It prints to stdout
out and in addition, also into the Synthesis log if one is currently
available via the EvolutionSyncClient.