packaging: updated rules for building .tar.gz and .deb
The old rules packaged just the command line tool. Now the .tar.gz
includes also the GUI (if enabled) in such a way that the user can
move the files into /usr. This is not recommended, but better than
not having those files at all.

The .deb packaging continues to use checkinstall, based on the theory
that with checkinstall it should be easier to generate both .deb and
.rpm. In practice this is a pretty ugly hack and probably should be
replaced. It also depends on a patched checkinstall (patches sent
upstream) which recognizes several additional options and works
under fakeroot.

It does not work for rpm, at least not without being root:
error: Could not open /usr/src/rpm/RPMS/x86_64/syncevolution-0.8.1+0.9+beta2+20090709-2.x86_64.rpm: Permission denied