packaging: some more dbus/interfaces fixes
Accidentally added the stamp- dummy file to the distro.

Also, noticed that the syncevo.xml file couldn't be
rebuilt from the original syncevo-full.xml. Not sure what
the original intention was, but now the files in git
are included in "make dist" and everything else has to
be built.

This leads to the xsltproc dependency which wasn't
in the README (see earlier commit) and not checked
for. Perhaps the intention was to avoid this
dependency when compiling a source .tar.gz.