.deb and .tar.gz binary dist: include server interoperability READMEs
All test/README.* files are included in the same directory as the
normal NEWS/COPYING/README files. It's no longer necessary that
users dig out a copy of these files from the source distro.

In Moblin users are not expected to read these files, so the
build via OBS will not be adapted.