SSL logging: added SYNCEVOLUTION_GNUTLS_DEBUG env variable to control libgnutls
The value of SYNCEVOLUTION_GNUTLS_DEBUG is directly passed to gnutls_global_set_log_level().
Use 10 to enable all logging. Log messages are written into the normal log with the
format "level %d: %s". GNUTLS is the line prefix, but that is not currently stored in .html.

To avoid a hard dependency on libgnutls (which we don't call directly), the relevant
functions are found via dlsym().