logging: don't use stderr/cerr (Bugzilla #4676)
Output via file descriptor 2 is filtered by our LogRedirect class.
This redirects errors into the log (if we have one) and suppresses
it otherwise (error output is translated to level DEVELOPER, which
is not printed on the console).

To get out own messages to the user reliably, we have to avoid
using stderr/cerr and print to the normal output channel instead.