Google + SSL: must use https, disable certificate checking
Google only supports https. Because of libsoup/gnutls limitiations
the certificate check has to be disabled (Bugzilla #4551). This
patch updates the default configuration for Google.

Users who update to a platform with working SSL must enable it
manually with such a configuration. Likewise, users who have
a configuration with SSL checking enabled and want to use it
on a platform with the too strict check must disable it in the
configuration.