GTK GUI: the 'your SyncML server account name' string was still shown in once case (Bugzilla #4273)
When editing an existing configuration the account name must be shown,
but not when that configuration is a template. I tried checking that
via "config->from_template" but that didn't have the desired effect,
so I check for "your SyncML server account name" explicitly now and
suppress that string.

Later on we need to duplicate that string in the GUI anyway so that
it can be translated.