calendar support: sanitize incoming EXDATEs (Bugzilla #4457)
Events which are relative to a time zone (detected based on DTSTART),
EXDATE values which have a time and are either in UTC (ScheduleWorld,
mobical.net) or floating (no timezone, not seen in the wild but might
occur), the EXDATE value is converted into the DTSTART timezone.

This is necessary for Evolution because it does not match the UTC
EXDATE with a recurrence in a time zone, but it is also a sensible
transformation in case that the time zone definition changes later on.