GTK-UI: crash due to use-after-free in error case (Bugzilla #4919)
The error callback frees its data pointer, but it gets called twice
with that pointer if the syncevo-dbus-server cannot be started.

As a quick fix (not the final one, which requires a deeper
understanding of memory ownership) this patch removes freeing
that data structure. Better leak some memory than crashing...