sync statistics: count items deleted during refresh-from-server (Bugzilla #3314)
The Synthesis Engine doesn't count these items because Synthesis wants
the engine to count only items deleted by the remote side. For our
statistics this looks a bit unrealistic (items before sync + added - deleted
does not add up to items after sync), so we now track items deleted
locally and use that number instead of the Synthesis statistics for deleted
items when we know that it is not complete.

In the command line, we show detailed statistics. In the GUI, it shows
up as part of the locally applied changes.

The new testComplexRefreshFromServerSemantic test covers this particular
aspect of the statistics.