XML configuration: always add mapping, using correct fieldlist name
The <fieldmap> is necessary to have access to individual fields.
Previously, only the "itemdata" variable was available to backends.
Now all fields can be read and written under their field names
(<automap>).

Because this only works if the correct fieldlist name is given,
this information now has to be provided as part of getSynthesisInfo().
Replaced list of strings with struct to allow future source-code
compatible API changes, like adding another field. Of course, this
will only work as long as the new fields are optional.