EvolutionContactSource: reimplemented logging via SyncSourceLogging
SyncEvolution 0.9 already supported logging of updated/inserted/deleted
contacts. This feature was removed when changing the backend API and
reimplemented using the common SyncSourceLogging class in this patch.

This is more efficient (no need to parse the item data) and provides
better results (prints "first middle last" name parts in that order,
with spaces inserted only where needed). In particular, the backslash
in "Doe\, John" is avoided.